+++ test/sampler_test_hdf.py
@@ -1,19 +1,16 @@
 import os
-import sys
 import shutil
 import numpy as np
 from scipy import stats
 
-src_path = os.path.dirname(os.path.abspath(__file__))
-sys.path.append(os.path.join(src_path, '..', 'src/mlmc'))
 from synth_simulation import SynthSimulation
 from synth_simulation_workspace import SynthSimulationWorkspace
-from sampler import Sampler
-from sample_storage_hdf import SampleStorageHDF
-from sampling_pool import OneProcessPool, ProcessPool
+from mlmc.sampler import Sampler
+from mlmc.sample_storage_hdf import SampleStorageHDF
+from mlmc.sampling_pool import OneProcessPool, ProcessPool
 from mlmc.moments import Legendre
-from quantity_estimate import QuantityEstimate
-import new_estimator
+from mlmc.quantity_estimate import QuantityEstimate
+import mlmc.new_estimator as new_estimator
 
 
 def sampler_hdf_test():
